requirements.txt
flask==2.1.1
werkzeug==2.1.1
tweepy==4.10.1
gunicorn==20.1.0
python-dotenv==0.19.2